Refactor: Improve component structure
