+++ src/utils/generic/formatRelativeDate.ts
@@ -1,0 +1,8 @@
+import { formatDistance } from "date-fns";
+
+const formatRelativeDate = (date: Date | string) => {
+  const ts = new Date(date);
+  return formatDistance(ts, new Date(), { addSuffix: true });
+};
+
+export default formatRelativeDate;
